[clean] 経験値瓶一括使用のコントローラを切り出す
